Supersedes #66.

This allows for phase tracking to only apply to the requested phases, rather than an 'all or nothing' like previously.

Will run the benchmarks later when I have time.