@PaintNinja 2025-04-04
Finally got a moment to run the benchmarks. Bare in mind that my hardware has changed, so these results are not directly comparable to other benchmarks I've shown in previous PRs.

Before:
```
Benchmark                                              Mode  Cnt     Score    Error  Units
BenchmarkModLauncher.Posting.Static.postStatic         avgt    5    10.791 ±  0.019  ns/op
BenchmarkModLauncher.Posting.Static.postStaticDozen    avgt    5    78.311 ±  0.368  ns/op
BenchmarkModLauncher.Posting.Static.postStaticHundred  avgt    5  1019.963 ± 41.705  ns/op
BenchmarkNoLoader.Posting.Static.postStatic            avgt    5    19.522 ±  0.102  ns/op
BenchmarkNoLoader.Posting.Static.postStaticDozen       avgt    5    93.828 ±  4.059  ns/op
BenchmarkNoLoader.Posting.Static.postStaticHundred     avgt    5  1738.962 ± 18.299  ns/op
```

After:
```
Benchmark                                              Mode  Cnt     Score    Error  Units
BenchmarkModLauncher.Posting.Static.postStatic         avgt    5     8.419 ±  0.012  ns/op
BenchmarkModLauncher.Posting.Static.postStaticDozen    avgt    5    78.679 ±  0.297  ns/op
BenchmarkModLauncher.Posting.Static.postStaticHundred  avgt    5   981.408 ± 84.251  ns/op
BenchmarkNoLoader.Posting.Static.postStatic            avgt    5    15.278 ±  0.025  ns/op
BenchmarkNoLoader.Posting.Static.postStaticDozen       avgt    5    89.780 ±  1.534  ns/op
BenchmarkNoLoader.Posting.Static.postStaticHundred     avgt    5  1654.612 ± 57.834  ns/op
```

So roughly 5%-20% faster when phase tracking is disabled or there are no registered listeners for the tracked phases.