@LexManos 2025-04-11
Not a fan of the default implementation being an exception throw. But guess that does allow for somewhat binary compatibility with older code that doesn't implement those methods. 

This looks fine. The performance metrics are dubious at best. But I can see the logic.
Tho it would break any consumer that attempts to perform logic based on phase. Not sure if we care/that exists.