@PaintNinja 2025-04-11
The default phase selection remains unchanged for back-compat. Disabling phase tracking disables all phases, and enabling phase tracking enables all phases, same as before.

Perf-wise you're welcome to run the benchmarks and compare yourself. The main benefit is for single listeners as obviously the difference going from posting 2 listeners to 1 listener is larger than going from 12 to 11 or 100 to 99.

If you're happy with this, please approve through GH and I'll merge.